Create custom loader and linker
This commit creates a custom loader and linker, which will ASLR the
binary on load. (The linker is a bash script that ensures that the
binary's layout is what the loader expects, and it also optimizes the
binary's section packing.)

The debug build still uses the system's LD and thus can run even if the
system doesn't support the loader's tricks (as its entry point is still
the beginning of .text, bypassing the loader entirely).