Dramatically improve efficiency of day 16, part 2
Instead of calculating the movement of both openers at once, calculate
them individually and then merge the best results, masked so that the
same valve isn't opened by both. This removes a factor of 2 at each
level of the decision tree in exchange for calculating 2^n decision
trees.
Also, assume that a result with some valves unopened will be the same
as the result if they had been disallowed: if the best score with all
valves allowed doesn't require opening valve QQ, then that score will
also be the best if QQ is disallowed. This reduces the number of
decision trees that need to be optimized by roughly two orders of
magnitude.